app/requirements.txt
fastapi
pydentic
pydantic[email]
sqlalchemy
uvicorn
celery
jinja2
bcrypt
python-dotenv
fastapi-jwt
PyJWT
python-multipart
